# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  login.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: aldenor
        type: string
    type: object
  login.TokenRequest:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  msresponse.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/msresponse.ErrorBody'
      message:
        example: Sucesso
        type: string
      ok:
        example: true
        type: boolean
      request_id:
        example: 7f8c2a9b-5f8a-4b0e-91e5-3b9c2d9a1234
        type: string
    type: object
  msresponse.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/msresponse.ErrorCode'
        example: 1
      description:
        example: Campo nm_obra é obrigatório
        type: string
    type: object
  msresponse.ErrorCode:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 500
    type: integer
    x-enum-varnames:
    - ErrorFormatoInvalido
    - ErrorTokenInvalido
    - ErrorNaoAutorizado
    - ErrorNaoEncontrado
    - ErrorValidacao
    - ErrorInterno
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://swagger.io
  description: This is a template server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io
  title: Server API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Valida usuário e senha e retorna um access token e um refresh token.
      parameters:
      - description: Dados de login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/login.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login realizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: Formato inválido ou campos obrigatórios ausentes
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "401":
          description: Usuário ou senha inválidos
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao gerar token
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Login do usuário
      tags:
      - Auth
...
  /auth/token/refresh:
    post:
      consumes:
      - application/json
      description: Valida um refresh token e retorna um novo access token.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/login.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login realizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: Formato inválido ou campos obrigatórios ausentes
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "401":
          description: Usuário ou senha inválidos
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao gerar token
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Renovar access token
      tags:
      - Auth
  /auth/token/verify:
    post:
      consumes:
      - application/json
      description: Verifica se um access token JWT é válido e retorna os dados principais
        das claims.
      parameters:
      - description: Token JWT
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/login.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login realizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: Formato inválido ou campos obrigatórios ausentes
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "401":
          description: Usuário ou senha inválidos
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Verificar token
      tags:
      - Auth
...
swagger: "2.0"